Hide filtered out refs
Indicate branches and tags in the left panel that are currently filtered
out and not visible in the revision grid.

The hidden refs are placed under a special node (in each participating
tree) called "[ Hidden ]".
NB: The label name contains a git-invalid char "[", so the node can't be
confused with genuine user-created refs (e.g. refs/heads/Hidden/branch).

Each node in the tree now has a new property `IsVisible` that can be used
as a conditional flag when implementing necessary logic.

Detection of whether a given ref is visible or not is done by checking
the revision grid, which is currently the source of truth.
Whenever the revision grid is refreshed, it raises an event that causes
the left panel to update its state, and show or hide nodes.